requirements.txt
streamlit
torch
torchvision
tqdm
scikit-learn
numpy
Pillow
